mmc: add debugfs hook to change host capabilities
This commit exposes the capability flags for a given host so that
flags can be changed at run time.  This is especially useful when
tuning a hardware bus so that specific modes and speeds can be locked
in.

Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Change-Id: I7ac5ab67ba7f7d8dd7c497b41f752c2643b0d586
Reviewed-on: http://gerrit.mot.com/720187
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>